requirements.txt
pytest==7.4.1
selene>=v2.0.0rc4
allure-pytest